Move still used classes
