+++ app/src/main/java/com/futsch1/medtimer/ActivityIntent.kt
@@ -2,8 +2,8 @@ package com.futsch1.medtimer
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
-import com.futsch1.medtimer.overview.customSnoozeDialog
-import com.futsch1.medtimer.overview.variableAmountDialog
+import com.futsch1.medtimer.new_overview.customSnoozeDialog
+import com.futsch1.medtimer.new_overview.variableAmountDialog
 
 fun dispatch(activity: AppCompatActivity, intent: Intent) {
     if (intent.action == "VARIABLE_AMOUNT") {
