+++ app/src/main/java/com/futsch1/medtimer/new_overview/ManualDose.kt
@@ -1,4 +1,4 @@
-package com.futsch1.medtimer.overview
+package com.futsch1.medtimer.new_overview
 
 import android.app.AlertDialog
 import android.content.Context
